+++ anylabeling/views/labeling/settings/schema.py
@@ -325,7 +325,7 @@ class SettingField:
 @lru_cache(maxsize=1)
 def load_template_config() -> dict[str, Any]:
     with pkg_resources.open_text(
-        anylabeling_configs, "xanylabeling_config.yaml"
+        anylabeling_configs, "jllabeling_config.yaml"
     ) as f:
         return yaml.safe_load(f)